feat: add native msvc build system and git lfs dependencies
Introduce native Windows MSVC (`cl.exe`) build orchestration via `make.bat` scripts across the entire repository with strict `-nostdlib++` parity, stage isolated SQLite3 prebuilt dependencies tracked by Git LFS, and standardize target naming across Makefiles.

Key architectural additions:
- Root and Subfolder MSVC Build Runners (`make.bat`): Orchestrate compilation targeting Microsoft Visual C++ (`cl.exe`) with `/GR-` (no RTTI), `/EHs-c-` (no exceptions), and `/link /NODEFAULTLIB:msvcprt.lib /NODEFAULTLIB:libcpmt.lib` to strictly prohibit linking against standard C++ runtime libraries.
- Git LFS Dependency Isolation (`deps/sqlite3/`, `.gitattributes`): Stage prebuilt SQLite 3.x headers (`sqlite3.h`, `sqlite3ext.h`), import library (`sqlite3.lib`), dynamic runtime libraries (`sqlite3.dll`, `libsqlite3-0.dll`), and CLI (`sqlite3.exe`) isolated from MinGW CRT headers to prevent MSVC intrinsic collisions.
- Compiler Warning & Strict Conformance Fixes:
  - `include/sqlite3_allocator.hpp`: Use `sqlite3_malloc64` in `sqlite_new_array` to eliminate 64-bit size truncation warnings.
  - `include/sqlite3_ext_state.h` & `include/sqlite3_ext_state.hpp`: Explicitly cast `strlen` lengths to prevent conversion warnings during registry entry allocation.
  - `tests/cpp_value/test_value.cpp`: Prevent MSVC compile-time divide-by-zero errors (`C2124`) during NaN test assertions via volatile floating-point expressions.
  - `tests/cpp_tvf/test_tvf_state.cpp`, `tests/cpp_udf/test_udf.cpp`, `tests/cpp_udf/test_udf_state.cpp`, `tests/cpp_vtab/test_vtab_state.cpp`: Add explicit integer casts from `as_int64()` arguments and `strlen` offsets.

Testing & Documentation:
- Root `Makefile`: Renamed and standardized C++ test targets to explicit `test-cpp-...` prefixes (`test-cpp-allocator`, `test-cpp-smart-ptr`, `test-cpp-udf`, `test-cpp-aggregate`, `test-cpp-statement`, `test-cpp-tvf`, `test-cpp-transaction`, `test-cpp-db`, `test-cpp-buffer`, `test-cpp-blob-stream`, `test-cpp-backup`, `test-cpp-vtab`, `test-cpp-extension`) and removed legacy alias targets.
- Verification Coverage: Verified 100% test pass rate across all test targets under both MSYS2 (`make test`) and native MSVC (`make.bat test`).
- Documentation: Updated `README.md` and `ARCHITECTURE.md` to document the dual build system architecture, MSVC command matrix, and zero-overhead runtime guarantees.

+++ include/sqlite3_ext_state.h
@@ -271,7 +271,7 @@
         StateType##_Entry *entry = (StateType##_Entry*) sqlite3_malloc(sizeof(StateType##_Entry)); \
         if (entry) { \
             /* Step 2: Deep copy the database path to serve as the registry key */ \
-            int path_len = strlen(db_path); \
+            int path_len = (int)strlen(db_path); \
             entry->db_path = (char*)sqlite3_malloc(path_len + 1); \
             if (entry->db_path) { \
                 memcpy(entry->db_path, db_path, path_len + 1); \
